content/media/test/test_metadata.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test returning metadata from media files with mozGetMetadata()</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var manager = new MediaTestManager;

function startTest(test, token) {
  var a = document.createElement('audio');
  a.token = token;
  manager.started(token);

  SimpleTest.waitForExplicitFinish();
  a.src = test.name;
  a.name = test.name;
  a.addEventListener('loadedmetadata', function() {
    tags = a.mozGetMetadata();
    ok(tags, "mozGetMetadata() failed");
    for (tag in tags) {
      is(tags[tag], test.tags[tag], "Tag '"+tag+"' doesn't match");
    }
    for (tag in test.tags) {
      is(tags[tag], test.tags[tag], "Tag '"+tag+"' doesn't match");
    }
    SimpleTest.finish();
  }, false);

}

manager.runTests(gMetadataTests, startTest);

</script>
</pre>
</body>
</html>
